+++ example/run_fractal_server.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+# set -e
 
 # Create and init db
 createdb fractal-example-test