+++ fractal_server/app/models/__init__.py
@@ -6,5 +6,4 @@ will not be picked up by alembic.
 from .linkusergroup import LinkUserGroup  # noqa: F401
 from .linkuserproject import LinkUserProjectV2  # noqa: F401
 from .security import *  # noqa
-from .user_settings import UserSettings  # noqa
 from .v2 import *  # noqa