+++ fractal_server/app/models/security.py
@@ -113,22 +113,9 @@ class UserOAuth(SQLModel, table=True):
         ondelete="RESTRICT",
     )
 
-    # TODO-2.17.1: update to `project_dir: str`
-    project_dir: str = Field(
-        sa_column=Column(
-            String,
-            server_default="/PLACEHOLDER",
-            nullable=False,
-        )
-    )
+    project_dir: str
     slurm_accounts: list[str] = Field(
         sa_column=Column(ARRAY(String), server_default="{}"),
-    )
-
-    # TODO-2.17.1: remove
-    user_settings_id: int | None = Field(
-        foreign_key="user_settings.id",
-        default=None,
     )