+++ tests/v2/fractal_tasks_mock/src/fractal_tasks_mock/dev/task_list.py
@@ -90,9 +90,10 @@ TASK_LIST = [
         executable="generic_task.py",
     ),
     NonParallelTask(
-        name="dummy_remove_images", executable="dummy_remove_images.py"
+        name="dummy_remove_images",
+        executable="dummy_remove_images.py",
     ),
-    NonParallelTask(
+    ConverterNonParallelTask(
         name="dummy_insert_single_image",
         executable="dummy_insert_single_image.py",
     ),